+++ source/CellMeasurer/index.js
@@ -1,4 +1,3 @@
 export default from './CellMeasurer'
 export CellMeasurer from './CellMeasurer'
 export defaultCellSizeCache from './defaultCellSizeCache'
-export uniformSizeCellSizeCache from './uniformSizeCellSizeCache'